Create .gitignore file and implement PEP guidelines - [#155735613]
